backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
pydantic[email]==2.9.2
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
python-multipart==0.0.19
alembic==1.14.0
python-dotenv==1.0.1
pymysql==1.1.1
psycopg[binary]==3.2.3
